Update to 18w07a

+++ src/main/java/org/lanternpowered/server/network/vanilla/message/type/play/MessagePlayOutChunkData.java
@@ -41,9 +41,9 @@ public final class MessagePlayOutChunkData implements Message {
     private final int z;
 
     private final Section[] sections;
-    @Nullable private final byte[] biomes;
+    @Nullable private final short[] biomes;
 
-    public MessagePlayOutChunkData(int x, int z, boolean skylight, Section[] sections, @Nullable byte[] biomes) {
+    public MessagePlayOutChunkData(int x, int z, boolean skylight, Section[] sections, @Nullable short[] biomes) {
         checkNotNull(sections, "sections");
         for (Section section : sections) {
             if (section != null) {
@@ -70,7 +70,7 @@ public final class MessagePlayOutChunkData implements Message {
     }
 
     @Nullable
-    public byte[] getBiomes() {
+    public short[] getBiomes() {
         return this.biomes;
     }
